# #74 provide integer (or binary) based EnumSet