--- a/.travis.yml
+++ b/.travis.yml
@@ -16,19 +16,19 @@ env:
 matrix:
   fast_finish: true
   include:
-    - php: 5.5
-      env:
-        - CODE_COVERAGE="1"
     - php: 5.6
       env:
-        - PHPDOC="1"
+        - CODE_COVERAGE="1"
     - php: 7.0
+      env:
+        - PHPDOC="1"
     - php: 7.1
       env:
         - CODE_COVERAGE="1"
     - php: nightly
+
+    # HHVM is no longer supported on Ubuntu Precise. Please consider using Trusty with `dist: trusty`.
     - php: hhvm
-      # HHVM is no longer supported on Ubuntu Precise. Please consider using Trusty with `dist: trusty`.
       dist: trusty
   allow_failures:
     - php: nightly