--- a/tests/MabeEnumTest/TestAsset/Enum31.php
+++ b/tests/MabeEnumTest/TestAsset/Enum31.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace MabeEnumTest\TestAsset;
+
+use MabeEnum\Enum;
+
+/**
+ * Unit tests for the class MabeEnum\Enum
+ *
+ * @link http://github.com/marc-mabe/php-enum for the canonical source repository
+ * @copyright Copyright (c) 2013 Marc Bennewitz
+ * @license http://github.com/marc-mabe/php-enum/blob/master/LICENSE.txt New BSD License
+ */
+class Enum31 extends Enum
+{
+    const ONE   = 1;
+    const TWO   = 2;
+    const THREE = 3;
+    const FOUR  = 4;
+    const FIVE  = 5;
+    const SIX   = 6;
+    const SEVEN = 7;
+    const EIGHT = 8;
+    const NINE  = 9;
+
+    const TEN       = 10;
+    const ELEVEN    = 11;
+    const TWELVE    = 12;
+    const THIRTEEN  = 13;
+    const FOURTEEN  = 14;
+    const FIVETEEN  = 15;
+    const SIXTEEN   = 16;
+    const SEVENTEEN = 17;
+    const EIGHTEEN  = 18;
+    const NINETEEN  = 19;
+    const TWENTY    = 20;
+
+    const TWENTYONE   = 21;
+    const TWENTYTWO   = 22;
+    const TWENTTHREE  = 23;
+    const TWENTYFOUR  = 24;
+    const TWENTYFIVE  = 25;
+    const TWENTYSIX   = 26;
+    const TWENTYSEVEN = 27;
+    const TWENTYEIGHT = 28;
+    const TWENTYNINE  = 29;
+
+    const THIRTY    = 30;
+    const THERTYONE = 31;
+}