@marc-mabe 2017-07-16
arg nothing to worry - I had a bug in my benchmarks - the PR is fixed now